:truck: Rename item type registries
