+++ src/main/app.ts
@@ -17,7 +17,7 @@ import { Notification } from 'electron';
 import { Backend, getBackend } from './backends';
 import { IMenuItem, IMenu, IMenuSettings, IAppSettings } from '../common';
 import { Settings, DeepReadonly } from './settings';
-import { ActionRegistry } from '../common/action-registry';
+import { ItemActionRegistry } from '../common/item-action-registry';
 
 declare const MAIN_WINDOW_PRELOAD_WEBPACK_ENTRY: string;
 declare const MAIN_WINDOW_WEBPACK_ENTRY: string;
@@ -342,7 +342,7 @@ export class KandoApp {
     // execute the action.
     ipcMain.on('select-item', (event, path) => {
       const execute = (item: DeepReadonly<IMenuItem>) => {
-        ActionRegistry.getInstance()
+        ItemActionRegistry.getInstance()
           .execute(item, this.backend)
           .catch((error) => {
             KandoApp.showError('Failed to execute action', error.message);
@@ -357,7 +357,7 @@ export class KandoApp {
         item = this.getMenuItemAtPath(this.lastMenu.nodes, path);
 
         // If the action is not delayed, we execute it immediately.
-        executeDelayed = ActionRegistry.getInstance().delayedExecution(item);
+        executeDelayed = ItemActionRegistry.getInstance().delayedExecution(item);
         if (!executeDelayed) {
           execute(item);
         }
