+++ src/renderer/editor/toolbar/trash-tab.ts
@@ -12,7 +12,7 @@ import { ToolbarItemDragger } from './toolbar-item-dragger';
 import { EventEmitter } from 'events';
 import { IMenu } from '../../../common';
 import { IEditorMenuItem } from '../common/editor-menu-item';
-import { ItemFactory } from '../../../common/item-factory';
+import { ItemTypeRegistry } from '../../../common/item-type-registry';
 import { IconThemeRegistry } from '../../../common/icon-theme-registry';
 
 /**
@@ -136,7 +136,7 @@ export class TrashTab extends EventEmitter {
 
       // If the item is a menu item, we need to extract the name and the icon.
       const item = thing as IEditorMenuItem;
-      const typeInfo = ItemFactory.getInstance().getItemType(item.type);
+      const typeInfo = ItemTypeRegistry.getInstance().getType(item.type);
       return {
         isMenu: false,
         name: item.name,
